+++ src/index.ts
@@ -1,63 +1,22 @@
-import { forIn } from "./ddzy/utility/object/forIn";
+import { fill } from "./ddzy/utility/array/fill";
 
-// ? 空对象
-interface IS1Props {
-};
-const s1 = {
-};
-forIn<IS1Props>(s1, (v, i, self) => {
-  console.log(v, i, self);
-});
 
-console.log('------------------------');
+// ? 空数组
+const s1: number[] = [];
+const p1 = fill<number>(s1, 100);
+console.log(p1);
 
-// ? 不应该提前退出
-interface IS2Props {
-  name: string,
-  age: number,
-  skill: string,
-};
-const s2: IS2Props = {
-  name: 'duanzhaoyang',
-  age: 21,
-  skill: 'program',
-};
-forIn(s2, (v) => {
-  console.log(v);
-});
+console.log('---------------------');
 
-console.log('------------------------------');
+// ? 非空数组
+const s2: number[] = [1, 2, 3, 4, 5];
+const p2 = fill<number>(s2, 200, 2, 8);
+console.log(p2);
 
-// ? 提前退出
-interface IS3Props {
-  city: string[];
-  position: {
-    x: number,
-    y: number,
-  };
-  continue: boolean;
-  nextOne: number;
-  nextTwo: number;
-};
-const s3: IS3Props = {
-  city: ['Dongguan', 'Foshan', 'Guangzhou'],
-  position: {
-    x: 100,
-    y: 200,
-  },
-  continue: false,
-  nextOne: 1,
-  nextTwo: 2,
-};
-forIn<IS3Props>(s3, function (v, i, origin) {
-  console.log({
-    this: this,
-    thisIsEqual: this === origin,
-    v,
-    i,
-  });
+console.log('-----------------------');
 
-  if (i === 'continue') {
-    return false;
-  }
-});
+// ? 原地操作
+const s3: any[] = [false, 0, '', NaN, true, function () { }];
+const p3 = fill<any>(s3, 'duanzhaoyang', 0);
+console.log(p3);
+console.log(p3 === s3);